# LeaderboardBackend/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "LeaderboardBackend",
    "version": "v1"
  },
  "paths": {
    "/Account/register": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Registers a new User.",
        "operationId": "register",
        "requestBody": {
          "description": "The `RegisterRequest` instance from which to register the `User`.",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/RegisterRequest"
                  }
                ],
                "description": "This request object is sent when a `User` is attempting to register."
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "The registration attempt was successfully received, and an email will\nbe sent to the provided address. If an account with that address does\nnot already exist, or if the account has not been confirmed yet, the\nemail will contain a link to confirm the account. Otherwise, the email\nwill inform the associated user that a registration attempt was made\nwith their address."
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "A `User` with the specified username already exists. The validation\nerror code `UsernameTaken` will be returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Logs a User in.",
        "operationId": "login",
        "requestBody": {
          "description": "The `LoginRequest` instance with which to perform the login.",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LoginRequest"
                  }
                ],
                "description": "This request object is sent when a `User` is attempting to log in."
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The `User` was logged in successfully. A `LoginResponse` is returned, containing a token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "The password given was incorrect, or no `User` could be found."
          },
          "403": {
            "description": "The associated `User` is banned."
          },
          "422": {
            "description": "The request contains errors.\nValidation error codes by property:\n- **Password**:\n  - **NotEmptyValidator**: No password was passed\n  - **PasswordFormat**: Invalid password format\n- **Email**:\n  - **NotEmptyValidator**: No email was passed\n  - **EmailValidator**: Invalid email format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Account/confirm": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Resends the account confirmation link.",
        "operationId": "resendConfirmationEmail",
        "responses": {
          "200": {
            "description": "A new confirmation link was generated."
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "The `User`'s account has already been confirmed."
          },
          "500": {
            "description": "The account recovery email failed to be created."
          }
        }
      }
    },
    "/Account/recover": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Sends an account recovery email.",
        "operationId": "sendRecoveryEmail",
        "requestBody": {
          "description": "The account recovery request.",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/RecoverAccountRequest"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "This endpoint returns 200 OK regardless of whether the email was sent successfully or not."
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Account/confirm/{id}": {
      "put": {
        "tags": [
          "Account"
        ],
        "summary": "Confirms a user account.",
        "operationId": "confirmAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The confirmation token.",
            "required": true,
            "schema": {
              "pattern": "^[a-zA-Z0-9-_]{22}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account was confirmed successfully."
          },
          "404": {
            "description": "The token provided was invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "the user's account was either already confirmed or banned."
          }
        }
      }
    },
    "/Account/recover/{id}": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Tests an account recovery token for validity.",
        "operationId": "testRecoveryToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The recovery token.",
            "required": true,
            "schema": {
              "pattern": "^[a-zA-Z0-9-_]{22}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The token provided is valid."
          },
          "404": {
            "description": "The token provided is invalid or expired, or the user is banned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Recover the user's account by resetting their password to a new value.",
        "operationId": "changePassword",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The recovery token.",
            "required": true,
            "schema": {
              "pattern": "^[a-zA-Z0-9-_]{22}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The password recovery request object.",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ChangePasswordRequest"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The user's password was reset successfully."
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The user is banned."
          },
          "404": {
            "description": "The token provided is invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The new password is the same as the user's existing password."
          },
          "422": {
            "description": "The request body contains errors.\nA **PasswordFormat** Validation error on the Password field indicates that the password format is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "required": [
          "password"
        ],
        "type": "object",
        "properties": {
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "description": "The `User`'s email address.",
            "format": "email",
            "example": "john.doe@example.com"
          },
          "password": {
            "minLength": 1,
            "type": "string",
            "description": "The `User`'s password. It:\n<ul><li>supports Unicode;</li><li>must be [8, 80] in length;</li><li>must have at least:</li><ul><li>one uppercase letter;</li><li>one lowercase letter; and</li><li>one number.</li></ul></ul>",
            "example": "P4ssword"
          }
        },
        "additionalProperties": false,
        "description": "This request object is sent when a `User` is attempting to log in."
      },
      "LoginResponse": {
        "required": [
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "minLength": 1,
            "type": "string",
            "description": "A JSON Web Token to authenticate and authorize queries with."
          }
        },
        "additionalProperties": false,
        "description": "This response object is received upon a successful log-in request."
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RecoverAccountRequest": {
        "required": [
          "email",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "minLength": 1,
            "type": "string",
            "description": "The user's name."
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "description": "The user's email address.",
            "format": "email"
          }
        },
        "additionalProperties": false
      },
      "RegisterRequest": {
        "required": [
          "email",
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "minLength": 1,
            "type": "string",
            "description": "The username of the `User`. It:\n<ul><li>must be [2, 25] in length;</li><li>must be made up of letters sandwiching zero or one of:</li><ul><li>hyphen;</li><li>underscore; or</li><li>apostrophe</li></ul></ul>\nUsernames are saved case-sensitively, but matched against case-insensitively.\nA `User` may not register with the name 'Cool' when another `User` with the name 'cool'\nexists.",
            "example": "J'on-Doe"
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "description": "The `User`'s email address.",
            "example": "john.doe@example.com"
          },
          "password": {
            "minLength": 1,
            "type": "string",
            "description": "The `User`'s password. It:\n<ul><li>supports Unicode;</li><li>must be [8, 80] in length;</li><li>must have at least:</li><ul><li>one uppercase letter;</li><li>one lowercase letter; and</li><li>one number.</li></ul></ul>",
            "example": "P4ssword"
          }
        },
        "additionalProperties": false,
        "description": "This request object is sent when a `User` is attempting to register."
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization using the Bearer scheme",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}